Add new routing and fix text colors on routes other than new landing-page
