+++ src/components/Footer/index.tsx
@@ -1,5 +1,6 @@
 import { useColorModeValue } from '@chakra-ui/color-mode';
 import { chakra, HStack, Text, VStack, Flex } from '@chakra-ui/react';
+import { useRouter } from 'next/router';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import {
@@ -13,12 +14,19 @@ import {
 
 function Footer() {
   const { t } = useTranslation();
+  const router = useRouter();
 
   const color = useColorModeValue('gray.600', '#fff');
 
   return (
     <chakra.footer borderTop="1px solid" borderColor="gray.200" py={10}>
-      <VStack color={color} mx="auto" maxW="7xl" px={4} spacing={5}>
+      <VStack
+        color={router.pathname === '/' ? color : 'gray.600'}
+        mx="auto"
+        maxW="7xl"
+        px={4}
+        spacing={5}
+      >
         <Flex alignItems="center" justifyContent="space-between" w="100%">
           <Text fontSize="sm">{t('madeBy')}</Text>
           <HStack as="ul" direction="row" spacing={4} listStyleType="none">
@@ -91,7 +99,10 @@ function Footer() {
           </HStack>
         </Flex>
         <VStack>
-          <Text color={color} fontSize="sm">
+          <Text
+            color={router.pathname === '/' ? color : 'gray.600'}
+            fontSize="sm"
+          >
             {t('hosting')}
           </Text>
           <chakra.a
