+++ src/layout/Page/index.tsx
@@ -12,7 +12,11 @@ function Page({ children }: { children?: object }) {
     'radial-gradient(26.46% 26.46% at 64.58% 8.69%, rgba(63, 70, 243, 0.18) 0%, rgba(63, 70, 243, 0) 100%), radial-gradient(43.06% 85.14% at 94.1% 7.91%, rgba(167, 170, 241, 0.03) 0%, rgba(238, 238, 252, 0.03) 100%), #0F0F1A',
   );
   return (
-    <Box height="100%" bg={router.pathname === '/landing-page' ? bg : ''}>
+    <Box
+      color={router.pathname !== '/' ? 'gray.800' : ''}
+      height="100%"
+      bg={router.pathname === '/' ? bg : ''}
+    >
       <Stack spacing={10}>
         <Nav />
         {children}
